requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
git+https://github.com/aliev/aioauth
sqlmodel==0.0.14
alembic==1.13.1
aiosqlite==0.19.0
PyCryptodome==3.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
orjson==3.9.10
jinja2==3.1.3
pydantic~=2.5.3
pydantic-settings==2.1.0
SQLAlchemy~=2.0.25
starlette~=0.35.1
pyrogram==2.0.106
tgcrypto==1.2.5
pytz~=2023.3.post1
APScheduler~=3.10.4
coloredlogs~=15.0.1
httpx==0.26.0
asyncmy==0.2.9
